Moved project to separate folder
