+++ ml_project/.hydra/config.yaml
@@ -1,8 +1,8 @@
-input_data_path: /data/heart_cleveland_upload.csv
-train_data_path: /data/train.csv
-test_data_path: /data/test.csv
-output_model_path: /models/model.pkl
-output_data_path: /data/predictions.csv
+input_data_path: data/heart_cleveland_upload.csv
+train_data_path: data/train.csv
+test_data_path: data/test.csv
+output_model_path: models/model.pkl
+output_data_path: data/predictions.csv
 splitting_params:
   val_size: 0.2
   random_state: 42
