+++ ml_project/.hydra/hydra.yaml
@@ -1,6 +1,6 @@
 hydra:
   run:
-    dir: ml_project
+    dir: .
   sweep:
     dir: multirun/${now:%Y-%m-%d}/${now:%H-%M-%S}
     subdir: ${hydra.job.num}
@@ -108,7 +108,7 @@ hydra:
   output_subdir: .hydra
   overrides:
     hydra:
-    - hydra.run.dir=ml_project
+    - hydra.run.dir=.
     task: []
   job:
     name: train
